+++ backoffice/src/Models/Entities/WreckMaterial.cs
@@ -14,10 +14,10 @@ public class WreckMaterial : BaseEntity
     public bool StoredAtSalvorAddress { get; set; } = false;
     public string? Description { get; set; } = string.Empty;
     public int Quantity { get; set; } = 1;
-    public float? Value { get; set; } = 0;
+    public double? Value { get; set; } = 0;
     
     public bool ValueKnown { get; set; } = false;
-    public float? ReceiverValuation { get; set; } = 0;
+    public double? ReceiverValuation { get; set; } = 0;
     public bool ValueConfirmed { get; set; } = false;
