+++ backoffice/src/Models/ViewModels/DroitView.cs
@@ -151,8 +151,8 @@ public class DroitView : BaseEntityView
 
 
     // Location
-    public float? Latitude { get; }
-    public float? Longitude { get; }
+    public double? Latitude { get; }
+    public double? Longitude { get; }
 
     [DisplayName("In UK Waters")]
     public bool InUkWaters { get; }
@@ -183,7 +183,7 @@ public class DroitView : BaseEntityView
 
     [DisplayName("Services Estimated Cost")]
 
-    public float? ServicesEstimatedCost { get; }
+    public double? ServicesEstimatedCost { get; }
 
     [DisplayName("MMO Licence Required")]
     public bool MMOLicenceRequired { get; }
@@ -192,7 +192,7 @@ public class DroitView : BaseEntityView
     public bool MMOLicenceProvided { get; }
 
     [DisplayName("Salvage Claim Awarded")]
-    public float SalvageClaimAwarded { get; }
+    public double SalvageClaimAwarded { get; }
 
     // Legacy fields
     public string? District { get; }
